+++ src/main/java/net/geforcemods/securitycraft/blockentities/LaserBlockBlockEntity.java
@@ -169,7 +169,7 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity implements INamed
 		}
 
 		excludedBEs.add(this);
-		createLinkedBlockAction(action, excludedBEs);
+		propagate(action, excludedBEs);
 	}
 
 	@Override
@@ -446,7 +446,7 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity implements INamed
 			for (ModuleType type : thatInsertedModules) {
 				bothInsertedModules.put(that.getModule(type).copy(), that.isModuleEnabled(type));
 				that.removeModule(type, false);
-				createLinkedBlockAction(new ILinkedAction.ModuleRemoved(type, false), that);
+				propagate(new ILinkedAction.ModuleRemoved(type, false), that);
 			}
 
 			readOptions(that.writeOptions(new CompoundNBT()));
@@ -458,9 +458,9 @@ public class LaserBlockBlockEntity extends LinkableBlockEntity implements INamed
 				ModuleType type = item.getModuleType();
 
 				insertModule(entry.getKey(), false);
-				createLinkedBlockAction(new ILinkedAction.ModuleInserted(module, item, false), this);
+				propagate(new ILinkedAction.ModuleInserted(module, item, false), this);
 				toggleModuleState(type, entry.getValue());
-				createLinkedBlockAction(new ILinkedAction.ModuleInserted(module, item, true), this);
+				propagate(new ILinkedAction.ModuleInserted(module, item, true), this);
 			}
 		}