+++ src/main/java/net/geforcemods/securitycraft/blocks/LaserFieldBlock.java
@@ -101,7 +101,7 @@ public class LaserFieldBlock extends OwnableBlock implements IOverlayDisplay, IW
 									laser.setLastToggleTime(System.currentTimeMillis());
 									level.setBlockAndUpdate(offsetPos, offsetState.cycle(LaserBlock.POWERED));
 									BlockUtils.updateIndirectNeighbors(level, offsetPos, SCContent.LASER_BLOCK.get());
-									laser.createLinkedBlockAction(new ILinkedAction.StateChanged<>(LaserBlock.POWERED, wasPowered, !wasPowered), laser);
+									laser.propagate(new ILinkedAction.StateChanged<>(LaserBlock.POWERED, wasPowered, !wasPowered), laser);
 
 									if (signalLength > 0)
 										level.getBlockTicks().scheduleTick(offsetPos, SCContent.LASER_BLOCK.get(), signalLength);